AuthContext & Form refactor
AuthContext - the context is created to pass the principal states of Authentication logic,  App -> Form, in Form the states change when a user log in or registers.
Form refactor - the Signin and Signup components are removed as a unique component and are integrated in Form component when they use the context to work.